Added link to output bulb, so image shown is linked to inputs
Changing the input values also updates the image of the lightbulb. Code is really really bad and really really messy, and the class I created should be split into two so that the output's functionality is in a second class, but it looks like it's functional for now, so I'll change it for the next sprint.