Put listeners in a loop and implemented rudimentary gate connections
Only did connection code for the gate class, Switch and Output remains unimplemented. Also, there currently exists a bug in where the line is rendered. Still need to also involve setting the color of the line based on the output value of the gate - and thus updating the color when inputs are changed and when connections are changed - among other things