PLEngine Code Integration
All PLEngine base functionality now in app, selectable from the main level select menu. Addtionally, the new UI was integrated for it, and minor adjustments made to LevelManager to account for the loss of previously valid tiles. LevelManager should now only ever reference tiles that remain valid in the app. NOTE: "restart" functionality is currently not working properly. It is likely this can be fixed in the final version as we'll have a set list of possible levels, making it easy to pass the parameters that will make the level function correctly to initiate the reset, but it is possible that we may instead have to switch the functionality to just a "clear board" method, which will function the same as a reset in all levels except timed ones