Next/Previous level button functionality
Clicking on the left/right arrows from the main level selection screen now scrolls the image card, loads the correct "level select" page, and will play the correct level. NOTE: level select accessed by clicking image card, as discussed in class; "play" button on main menu remains for now, but is obsolete and is to be removed in later commits. Additionally, clicking the "hard" or "easy" buttons in level select will change the colors of the buttons.
IN PROGRESS: working to fix functionality of quit level/beating a level, so that doing so will bring you to the correct highlighted card of the main menu (either the current level's card if "quit" was chosen or level was lost, or the next card if game was successfully completed)